Make hlint tests work again
We also do this in a two stage process: we use runCommand to find Haskell repos,
import the result and generate a test script for each. This way we can run those
test scripts over and over without having to lookup repos again.

To force the repos to be looked up, we just rebuild the test command.